+++ src/components/ToggleMenu/ToggleMenu.js
@@ -22,12 +22,42 @@ const MenuWrapper = styled.div`
   min-width: 100%;
 `;
 
+const getDefaultDirection = position => {
+  switch (position) {
+    case 'left':
+    case 'right':
+      return 'bottom';
+    default:
+      return 'right';
+  }
+};
+
+const getDerivedPosition = position => ({
+  place: position,
+  direction: getDefaultDirection(position),
+});
+
+const getOppositeDirection = direction => {
+  switch (direction) {
+    case 'top':
+      return 'bottom';
+    case 'bottom':
+      return 'top';
+    case 'left':
+      return 'right';
+    case 'right':
+      return 'left';
+    default:
+      return direction;
+  }
+};
+
 class ToggleMenu extends Component {
   constructor(props) {
     super(props);
     this.state = {
       open: false,
-      position: props.position,
+      position: getDerivedPosition(props.position),
     };
   }
 
@@ -61,46 +91,40 @@ class ToggleMenu extends Component {
     this.removeDocumentEventListeners();
   }
 
-  getPositionValues = position => {
+  getPositionValues = ({ place, direction }) => {
     const { menuOffset } = this.props;
     const positions = {
       top: {
         bottom: '100%',
-        left: '0',
+        [getOppositeDirection(direction)]: '0',
         margin: `0 0 ${menuOffset} 0`,
       },
       bottom: {
         top: '100%',
-        left: '0',
+        [getOppositeDirection(direction)]: '0',
         margin: `${menuOffset} 0 0 0`,
       },
       left: {
         right: '100%',
-        top: '0',
+        [getOppositeDirection(direction)]: '0',
         margin: `0 ${menuOffset} 0 0`,
       },
       right: {
         left: '100%',
-        top: '0',
+        [getOppositeDirection(direction)]: '0',
         margin: `0 0 0 ${menuOffset}`,
       },
     };
-    return positions[position];
+    return positions[place];
   };
 
-  getOppositePosition = position => {
-    switch (position) {
-      case 'top':
-        return 'bottom';
-      case 'bottom':
-        return 'top';
-      case 'left':
-        return 'right';
-      case 'right':
-        return 'left';
-      default:
-        return position;
-    }
+  getPreferredPosition = withinBounds => {
+    const { position } = this.state;
+    const { place, direction } = position;
+    return {
+      place: withinBounds[place] ? place : getOppositeDirection(place),
+      direction: withinBounds[direction] ? direction : getOppositeDirection(direction),
+    };
   };
 
   addDocumentEventListeners = () => {
@@ -132,19 +156,20 @@ class ToggleMenu extends Component {
 
   handleOutOfBounds = (restoreOriginalPosition = false) => {
     const withinBounds = this.isMenuInViewport();
-    const outOfBoundsSide = Object.keys(withinBounds).find(key => withinBounds[key] === false);
-    if (!outOfBoundsSide) {
-      // not out of bounds.
+    const allWithin = Object.keys(withinBounds).filter(key => !withinBounds[key]).length === 0;
+    if (allWithin) {
       return;
     }
 
     if (!restoreOriginalPosition) {
-      this.setState({ position: this.getOppositePosition(outOfBoundsSide) }, () => {
+      this.setState({ position: this.getPreferredPosition(withinBounds) }, () => {
         this.handleOutOfBounds(true);
       });
     } else {
       // restores the original position if both opposite sides are out of bounds.
-      this.setState({ position: this.props.position });
+      this.setState({
+        position: getDerivedPosition(this.props.position),
+      });
     }
   };
 
@@ -162,7 +187,7 @@ class ToggleMenu extends Component {
       return;
     }
 
-    this.setState({ open: true, position: this.props.position }, () => {
+    this.setState({ open: true, position: getDerivedPosition(this.props.position) }, () => {
       if (this.props.closeOnOutsideClick) {
         this.addDocumentEventListeners();
       }
@@ -239,7 +264,8 @@ ToggleMenu.propTypes = {
   onOpen: PropTypes.func,
   onClose: PropTypes.func,
   closeOnOutsideClick: PropTypes.bool,
-  menuOffset: PropTypes.string, // distance between the menu and the trigger
+  /** Distance between the menu and the trigger. */
+  menuOffset: PropTypes.string,
   fluid: PropTypes.bool,
   open: PropTypes.bool,
   persist: PropTypes.bool,
@@ -252,7 +278,7 @@ ToggleMenu.defaultProps = {
   onOpen: () => null,
   onClose: () => null,
   closeOnOutsideClick: true,
-  menuOffset: '5px',
+  menuOffset: '0.25rem',
   fluid: false,
   persist: false,
 };